database/requirements.txt
mysql-connector-python==8.1.0
SQLAlchemy==2.0.23
pandas==2.1.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9 
pyodbc==4.0.39